requirements.txt
# VoiceType - Voice to Text Application
# ======================================
# Install: pip install -r requirements.txt

# GUI Framework
customtkinter>=5.2.0      # Modern Tkinter widgets
Pillow>=10.0.0            # Image processing

# Audio Processing
sounddevice>=0.4.6        # Audio recording (recommended)
numpy>=1.24.0             # Numerical operations
noisereduce>=3.0.0        # Noise reduction for better audio quality
scipy>=1.10.0             # Signal processing for audio
torch>=2.0.0              # PyTorch for Silero VAD (CPU version)
torchaudio>=2.0.0         # Audio processing with PyTorch

# OpenAI API
openai>=1.3.0             # OpenAI SDK for Whisper & GPT
httpx>=0.25.0             # HTTP client

# Hotkeys & Keyboard Control
pynput>=1.7.6             # Global hotkey listener
keyboard>=0.13.5          # Keyboard control (backup)

# Clipboard & Text Injection
pyperclip>=1.8.2          # Cross-platform clipboard
pyautogui>=0.9.54         # GUI automation for paste

# System Tray (Optional)
pystray>=0.19.4           # System tray icon

# Windows API (Windows only)
pywin32>=306; sys_platform == 'win32'

# Environment Variables
python-dotenv>=1.0.0      # Load .env files

# Build (Optional)
# pyinstaller>=6.0.0      # Create standalone executable
